+++ R/ciu.R
@@ -156,6 +156,27 @@ ciu.plot <- function(ciu, instance, ind.input, ind.output, in.min.max.limits=NUL
   ciu$plot.ciu (instance, ind.input, ind.output, in.min.max.limits, n.points, main, xlab, ylab, ylim, ...)
 }
 
+#' ciu.ggplot
+#'
+#' Function for plotting out the effect of changing values of one input on one output.
+#'
+#' @inheritParams ciu.plot
+#' @param illustrate.CIU Include illustration of CIU Cmin, Cmax, neutral.CU. Default is FALSE
+#' @param neutral.CU Value of neutral.CU. Default is 0.5.
+#' @param CIU.illustration.colours Colours to use for illustrating CIU.
+#' Default is red, orange, green.
+#'
+#' @return ggplot object.
+#' @export
+#' @author Kary Främling
+ciu.ggplot <- function(ciu, instance, ind.input=1, ind.output=1, in.min.max.limits=NULL, n.points=40, main=NULL, xlab=NULL, ylab=NULL,
+                       ylim=NULL, illustrate.CIU=FALSE, neutral.CU=0.5, CIU.illustration.colours=c("red", "orange", "green")) {
+  if ( inherits(ciu, "ciu") )
+    ciu <- ciu.to.CIU(ciu)
+  ciu$ggplot.ciu(instance, ind.input, ind.output, in.min.max.limits, n.points, main, xlab, ylab,
+                 ylim, illustrate.CIU, neutral.CU, CIU.illustration.colours)
+}
+
 #' ciu.plot.3D
 #'
 #' Function for 3D plotting the effect of changing values of two inputs on one output.